Replace GSL matrix inversion with Scheme code for same.
Currently the ICP code has a segment violation every once in a while. I'm not sure where or why (if I was, I'd fix it!) I'm trying to remove dependencies on funky FFI code to narrow down the bug / remove possible bug causers. This doesn't fix it, but it is a good improvement anyway.